pdf2markdown/requirements.txt
pdfplumber
pdfminer.six
tqdm
python-dotenv
requests<3.0.0
urllib3<2.0.0
openai
tenacity
ollama 